-added a controller method to return activity type info -modified models and diagram to support activity type name/fore/back/icon

+++ d360.model/WorkflowContext.cs
@@ -212,7 +212,7 @@ namespace d360.model
         {
             var itemStep = getWorkflowItemStep(itemStepID);
 
-            var activityType = ActivityTypes.SingleOrDefault(i => i.Value.ID == itemStep.Step.ActivityType);
+            var activityType = ActivityTypes.SingleOrDefault(i => i.Value.ID == (int)itemStep.Step.ActivityType);
             if (activityType == null)
                 throw new ApplicationException($"Item Step does not correspond to any known activity type of {itemStep.Step.ActivityType}.");
